Refactor shared SEP24 & SEP6 use cases with the following changes:
- Avoid inner functions.
- Rename function as infinitive.
- Add a provider to instantiate the correct protocol.
