+++ nodes/Anclap/actions/info/transactions/execute.ts
@@ -1,12 +1,12 @@
 import { IExecuteFunctions } from 'n8n-workflow';
 import AnclapCredentials from '../../../transport/AnclapCredentials';
-import SEP24 from '../../../transport/SEP24';
-import SEP6 from '../../../transport/SEP6';
-import { Protocol, TransactionType } from '../../../transport/types';
+import { TransactionType } from '../../../transport/types';
 import TransactionsRequest from '../../../transport/requests/TransactionsRequest/TransactionsRequest';
 import ITransactionsRequest from '../../../transport/requests/TransactionsRequest/ITransactionsRequest';
+import getProtocolProvider from '../../../transport/providers/protocolProvider';
+import { Protocol } from '../../../transport/enums/protocol';
 
-export async function transactions(this: IExecuteFunctions) {
+export async function getTransactions(this: IExecuteFunctions) {
 	const anclapCredentials = new AnclapCredentials(await this.getCredentials('anclapApi'));
 	const token = this.getNodeParameter('token', 0) as string;
 	const protocol = this.getNodeParameter('protocol', 0) as Protocol;
@@ -35,19 +35,8 @@ export async function transactions(this: IExecuteFunctions) {
 	} else {
 		transactionsRequest = new TransactionsRequest({ assetCode });
 	}
-	async function getSep24Transactions() {
-		const sep24 = new SEP24(anclapCredentials, token);
-		return await sep24.getTransactions(transactionsRequest);
-	}
 
-	async function getSep6Transactions() {
-		const sep6 = new SEP6(anclapCredentials, token);
-		return await sep6.getTransactions(transactionsRequest);
-	}
+	const transactionsProvider = getProtocolProvider(anclapCredentials, token, protocol);
 
-	if (protocol === 'sep24') {
-		return await getSep24Transactions();
-	} else {
-		return await getSep6Transactions();
-	}
+	return await transactionsProvider.getTransactions(transactionsRequest);
 }
