+++ nodes/Anclap/actions/info/transfer-server/execute.ts
@@ -1,40 +1,22 @@
 import { IExecuteFunctions } from 'n8n-workflow';
 import AnclapCredentials from '../../../transport/AnclapCredentials';
-import SEP6 from '../../../transport/SEP6';
-import SEP24 from '../../../transport/SEP24';
 import TransferServerRequest from '../../../transport/requests/TransferServerRequest/TransferServerRequest';
-import { Protocol } from '../../../transport/types';
+import getProtocolProvider from '../../../transport/providers/protocolProvider';
+import { Protocol } from '../../../transport/enums/protocol';
 
-export async function transferServer(this: IExecuteFunctions) {
+export async function getTransferServer(this: IExecuteFunctions) {
 	const anclapCredentials = new AnclapCredentials(await this.getCredentials('anclapApi'));
 	const protocol = this.getNodeParameter('protocol', 0) as Protocol;
 
-	const getTransferServerInfo = async () => {
-		const showOptionalValues = this.getNodeParameter('showOptionalValues', 0) as boolean;
+	const showOptionalValues = this.getNodeParameter('showOptionalValues', 0) as boolean;
 
-		let transferServerRequest: TransferServerRequest;
-		if (showOptionalValues) {
-			const lang = this?.getNodeParameter('lang', 0) as string;
+	let transferServerRequest: TransferServerRequest = {};
+	if (showOptionalValues) {
+		const lang = this?.getNodeParameter('lang', 0) as string;
 
-			transferServerRequest = new TransferServerRequest({ lang });
-		}
+		transferServerRequest = new TransferServerRequest({ lang });
+	}
+	const transferServerProvider = getProtocolProvider(anclapCredentials, '', protocol);
 
-		async function getSep24Info() {
-			const sep24 = new SEP24(anclapCredentials, '');
-			return await sep24.getInfo(transferServerRequest);
-		}
-
-		async function getSep6Info() {
-			const sep6 = new SEP6(anclapCredentials, '');
-			return await sep6.getInfo(transferServerRequest);
-		}
-
-		if (protocol === 'sep24') {
-			return await getSep24Info();
-		} else {
-			return await getSep6Info();
-		}
-	};
-
-	return await getTransferServerInfo();
+	return transferServerRequest ? await transferServerProvider.getInfo(transferServerRequest): await transferServerProvider.getInfo();
 }
